+++ client/src/apollo-recipes/panels/example-panel1.js
@@ -1,8 +1,8 @@
 import gql from 'graphql-tag';
-import { Panel } from '../panel_components.js';
+import { InfographicPanel } from '../panel_components.js';
 
 const Component = ({ data: { programs } }) => (
-  <Panel title="example panel 2">
+  <InfographicPanel title="example panel 2">
     <ul>
       {_.map(programs, ({name}) =>
         <li key={name}>
@@ -10,7 +10,7 @@ const Component = ({ data: { programs } }) => (
         </li>
       )}
     </ul>
-  </Panel>
+  </InfographicPanel>
 );
