+++ client/src/panels/InfographicPanel.yaml
