+++ client/src/panels/intro_graphs/gov_related.js
@@ -3,7 +3,7 @@ import {
   declare_panel,
   create_text_maker,
   TM,
-  Panel,
+  InfographicPanel,
 } from "../shared.js";
 
 const text_maker = create_text_maker(text);
@@ -19,14 +19,14 @@ export const declare_gov_related_info_panel = () => declare_panel({
     calculate: _.constant(true),
     render(){
       return (
-        <Panel title={text_maker("gov_related_info_title")}>
+        <InfographicPanel title={text_maker("gov_related_info_title")}>
           <div 
             className="medium_panel_text"
             style={{lineHeight: "40px"}}
           >
             <TM tmf={text_maker} k="gov_related_info_text" />
           </div>
-        </Panel>
+        </InfographicPanel>
       );
     },
   }),
